Updated files
A lot of changes in this one, mainly in Game. A lot of it is methods that haven't been implemented yet, but it should make sense in an abstract way. Running the Game class has an error when reading user input. It works for the number of players, but it threw errors for making a move. It's to do with waiting for the user to input something.